@@ -1,70 +1,60 @@
-<?php namespace YAAP\Theme;
+<?php
+
+namespace YAAP\Theme;
 
 use Illuminate\Container\Container;
+use Illuminate\Contracts\Translation\Loader;
+use Illuminate\Support\Arr;
 use Illuminate\View\ViewFinderInterface;
 use YAAP\Theme\Exceptions\ThemeException;
-use Illuminate\Support\Arr;
-use Illuminate\Contracts\Translation\Loader;
 
 /**
- * Class Theme
- * @package YAAP\Theme
+ * Class Theme.
  */
 class Theme
 {
-
     /**
-     * The IoC Container
+     * The IoC Container.
      *
      * @var Container
      */
     protected $app;
 
     /**
-     * Default view finder
-     *
+     * Default view finder.
      */
-
     protected $finder;
 
     /**
-     * Default locale loader
-     *
+     * Default locale loader.
      */
-
     protected $localeLoader;
 
     /**
-     * theme config
-     *
+     * theme config.
      */
     protected $config;
 
     /**
-     * parent themes
-     *
+     * parent themes.
      */
     protected $parents;
 
     /**
-     *  cache for paths
+     *  cache for paths.
      */
     protected $cache;
 
     /**
-     *  current theme
+     *  current theme.
      */
     protected $theme;
 
     /**
-     * Build a new Theme manager
-     *
-     * @param Container $app
-     * @param ViewFinderInterface $finder
+     * Build a new Theme manager.
      */
     public function __construct(Container $app, ViewFinderInterface $finder, Loader $localeLoader)
     {
-
         $this->app = $app;
 
         $this->finder = $finder;
@@ -72,31 +62,32 @@ class Theme
         $this->localeLoader = $localeLoader;
     }
 
-
     /**
+     * Initialize a theme by name.
      *
-     * Initialize a theme by name
-     * @param $theme
      * @throws ThemeException
      */
-    public function init($theme)
+    public function init($theme): void
     {
-        if (empty($theme)) throw new ThemeException('Theme name should not be empty');
+        if (empty($theme)) {
+            throw new ThemeException('Theme name should not be empty');
+        }
 
         $this->theme = $theme;
 
         // read theme path
         $path = $this->app['config']->get('theme.path', base_path('themes'));
 
-        //init config
-        $this->config = $this->_readConfig( $path . '/' . $theme . '/config.php');
+        // init config
+        $this->config = $this->_readConfig($path . '/' . $theme . '/config.php');
 
         // theme parents
-        $this->parents = array();
+        $this->parents = [];
 
         while (!empty($theme)) {
-
-            if (!is_dir($path . '/' . $theme)) throw new ThemeException('Theme '.$theme.' not found.');
+            if (!is_dir($path . '/' . $theme)) {
+                throw new ThemeException('Theme ' . $theme . ' not found.');
+            }
 
             // add theme's root folder
             $this->finder->addLocation($path . '/' . $theme);
@@ -115,7 +106,6 @@ class Theme
         }
 
         $this->localeLoader->addNamespace($this->theme, $path . '/' . $this->theme . '/' . 'lang');
-
     }
 
     /**
@@ -128,29 +118,27 @@ class Theme
         // read theme path
         $path = $this->app['config']->get('theme.path', base_path('themes'));
 
-        if (file_exists($path))
-        {
+        if (file_exists($path)) {
             $dir_list = dir($path);
             while (false !== ($entry = $dir_list->read())) {
-                if (file_exists($path . '/' . $entry . '/' . 'config.php'))
+                if (file_exists($path . '/' . $entry . '/' . 'config.php')) {
                     $list[] = $entry;
+                }
             }
         }
 
         return $list;
-
     }
 
     /**
-     * @param $path
      * @return array|mixed
      */
-    private function _readConfig($path){
-        if (file_exists($path))
-                return include($path);
+    private function _readConfig($path)
+    {
+        if (file_exists($path)) {
+            return include $path;
+        }
 
-        return array();
+        return [];
     }
-
 }
-
