@@ -1,7 +1,8 @@
 <?php
 
-return [
+declare(strict_types=1);
 
+return [
     /*
     |--------------------------------------------------------------------------
     | Theme name
@@ -23,5 +24,4 @@ return [
     */
 
     'inherit' => null,
-
 ];
